+++ EverythingSDK/include/Everything_IPC.h
