+++ EverythingSDK/src/Everything.c
@@ -35,7 +35,7 @@
 
 // include
 #include "../include/Everything.h"
-#include "../ipc/Everything_IPC.h"
+#include "../include/Everything_IPC.h"
 
 // return copydata code
 #define _EVERYTHING_COPYDATA_QUERYREPLY		0
